feat: Add onKeyDown event handler for password input in OnboardingStep2 component
This commit adds an onKeyDown event handler for the password input in the OnboardingStep2 component. When the Enter key is pressed, the handleVerify function is called to verify the API key. This change improves the user experience by allowing them to press Enter to submit the API key instead of having to click a button.